header files needed to change return type of getCurrentMillis to unsigned long long

+++ examples/c/util.h
@@ -15,7 +15,7 @@
 /**
  * @return the number of milliseconds elapsed since the Epoch.
  */
-unsigned long
+unsigned long long
 getCurrentMillis (void);
 
 /**
